assembler refactor
