+++ vyper/evm/assembler/__init__.py
@@ -1,0 +1,4 @@
+from vyper.evm.assembler.core import assembly_to_evm, resolve_symbols
+from vyper.evm.assembler.optimizer import optimize_assembly
+
+__all__ = ["assembly_to_evm", "resolve_symbols", "optimize_assembly"]
