+++ vyper/ir/compile_ir.py
@@ -2,184 +2,21 @@ from __future__ import annotations
 
 import contextlib
 import copy
-from dataclasses import dataclass
-from typing import Any, Optional, TypeVar
+from typing import Any, Optional
 
+from build.lib.vyper.ir.compile_ir import DataHeader
 import cbor2
 
 from vyper.codegen.ir_node import IRnode
 from vyper.compiler.settings import OptimizationLevel
-from vyper.evm.opcodes import get_opcodes, version_check
+from vyper.evm.assembler.core import assembly_to_evm, get_data_segment_lengths
+from vyper.evm.assembler.instructions import CONST, DATA_ITEM, JUMP, JUMPI, PUSH, PUSH_N, PUSH_OFST, PUSHLABEL, AssemblyInstruction, TaggedInstruction, mkdebug
+from vyper.evm.assembler.optimizer import optimize_assembly
+from vyper.evm.assembler.symbols import CONSTREF, Label
+from vyper.evm.opcodes import get_opcodes
 from vyper.exceptions import CodegenPanic, CompilerPanic
-from vyper.ir.optimizer import COMMUTATIVE_OPS
-from vyper.utils import MemoryPositions, OrderedSet
+from vyper.utils import MemoryPositions
 from vyper.version import version_tuple
-
-PUSH_OFFSET = 0x5F
-DUP_OFFSET = 0x7F
-SWAP_OFFSET = 0x8F
-
-
-def num_to_bytearray(x):
-    o = []
-    while x > 0:
-        o.insert(0, x % 256)
-        x //= 256
-    return o
-
-
-def PUSH(x):
-    bs = num_to_bytearray(x)
-    # starting in shanghai, can do push0 directly with no immediates
-    if len(bs) == 0 and not version_check(begin="shanghai"):
-        bs = [0]
-    return [f"PUSH{len(bs)}"] + bs
-
-
-# push an exact number of bytes
-def PUSH_N(x, n):
-    o = []
-    for _i in range(n):
-        o.insert(0, x % 256)
-        x //= 256
-    assert x == 0
-    return [f"PUSH{len(o)}"] + o
-
-
-#####################################
-# assembly data structures and utils
-#####################################
-
-
-class Label:
-    def __init__(self, label: str):
-        assert isinstance(label, str)
-        self.label = label
-
-    def __repr__(self):
-        return f"LABEL {self.label}"
-
-    def __eq__(self, other):
-        if not isinstance(other, Label):
-            return False
-        return self.label == other.label
-
-    def __hash__(self):
-        return hash(self.label)
-
-
-@dataclass
-class DataHeader:
-    label: Label
-
-    def __repr__(self):
-        return f"DATA {self.label.label}"
-
-
-# this could be fused with Label, the only difference is if
-# it gets looked up from const_map or symbol_map.
-class CONSTREF:
-    def __init__(self, label: str):
-        assert isinstance(label, str)
-        self.label = label
-
-    def __repr__(self):
-        return f"CONSTREF {self.label}"
-
-    def __eq__(self, other):
-        if not isinstance(other, CONSTREF):
-            return False
-        return self.label == other.label
-
-    def __hash__(self):
-        return hash(self.label)
-
-
-class CONST:
-    def __init__(self, name: str, value: int):
-        assert isinstance(name, str)
-        assert isinstance(value, int)
-        self.name = name
-        self.value = value
-
-    def __repr__(self):
-        return f"CONST {self.name} {self.value}"
-
-    def __eq__(self, other):
-        if not isinstance(other, CONST):
-            return False
-        return self.name == other.name and self.value == other.value
-
-
-class PUSHLABEL:
-    def __init__(self, label: Label):
-        assert isinstance(label, Label), label
-        self.label = label
-
-    def __repr__(self):
-        return f"PUSHLABEL {self.label.label}"
-
-    def __eq__(self, other):
-        if not isinstance(other, PUSHLABEL):
-            return False
-        return self.label == other.label
-
-    def __hash__(self):
-        return hash(self.label)
-
-
-# push the result of an addition (which might be resolvable at compile-time)
-class PUSH_OFST:
-    def __init__(self, label: Label | CONSTREF, ofst: int):
-        # label can be Label or CONSTREF
-        assert isinstance(label, (Label, CONSTREF))
-        self.label = label
-        self.ofst = ofst
-
-    def __repr__(self):
-        label = self.label
-        if isinstance(label, Label):
-            label = label.label  # str
-        return f"PUSH_OFST({label}, {self.ofst})"
-
-    def __eq__(self, other):
-        if not isinstance(other, PUSH_OFST):
-            return False
-        return self.label == other.label and self.ofst == other.ofst
-
-    def __hash__(self):
-        return hash((self.label, self.ofst))
-
-
-class DATA_ITEM:
-    def __init__(self, item: bytes | Label):
-        self.data = item
-
-    def __repr__(self):
-        if isinstance(self.data, bytes):
-            return f"DATABYTES {self.data.hex()}"
-        elif isinstance(self.data, Label):
-            return f"DATALABEL {self.data.label}"
-
-
-def JUMP(label: Label):
-    return [PUSHLABEL(label), "JUMP"]
-
-
-def JUMPI(label: Label):
-    return [PUSHLABEL(label), "JUMPI"]
-
-
-def mkdebug(pc_debugger, ast_source):
-    # compile debug instructions
-    # (this is dead code -- CMC 2025-05-08)
-    i = TaggedInstruction("DEBUG", ast_source)
-    i.pc_debugger = pc_debugger
-    return [i]
-
-
-def is_label(i):
-    return isinstance(i, Label)
 
 
 def generate_cbor_metadata(
@@ -226,19 +63,6 @@ def _runtime_code_offsets(ctor_mem_size, runtime_codelen):
     return runtime_code_start, runtime_code_end
 
 
-# Calculate the size of PUSH instruction
-def calc_push_size(val: int):
-    # stupid implementation. this is "slow", but its correctness is
-    # obvious verify, as opposed to
-    # ```
-    # (val.bit_length() + 7) // 8
-    #    + (1
-    #         if (val > 0 or version_check(begin="shanghai"))
-    #      else 0)
-    # ```
-    return len(PUSH(val))
-
-
 # temporary optimization to handle stack items for return sequences
 # like `return return_ofst return_len`. this is kind of brittle because
 # it assumes the arguments are already on the stack, to be replaced
@@ -276,19 +100,6 @@ def _rewrite_return_sequences(ir_node, label_params=None):
     for t in args:
         _rewrite_return_sequences(t, label_params)
 
-
-# a string (assembly instruction) but with additional metadata from the source code
-class TaggedInstruction(str):
-    def __new__(cls, sstr, *args, **kwargs):
-        return super().__new__(cls, sstr)
-
-    def __init__(self, sstr, ast_source=None, error_msg=None):
-        self.error_msg = error_msg
-        self.pc_debugger = False
-
-        self.ast_source = ast_source
-
-
 ##############################
 # IRnode to assembly
 ##############################
@@ -319,14 +130,6 @@ def compile_to_assembly(
     if optimize != OptimizationLevel.NONE:
         optimize_assembly(res)
     return res
-
-
-# TODO: move all these assembly data structures to own module, like
-# vyper.evm.assembly
-AssemblyInstruction = (
-    str | TaggedInstruction | int | PUSHLABEL | Label | PUSH_OFST | DATA_ITEM | DataHeader | CONST
-)
-
 
 class _IRnodeLowerer:
     # map from variable names to height in stack
@@ -956,543 +759,3 @@ class _IRnodeLowerer:
         return JUMPI(self.global_revert_label)
 
 
-##############################
-# assembly to evm utils
-##############################
-
-
-def note_line_num(line_number_map, pc, item):
-    # Record AST attached to pc
-    if isinstance(item, TaggedInstruction):
-        if (ast_node := item.ast_source) is not None:
-            ast_node = ast_node.get_original_node()
-            if hasattr(ast_node, "node_id"):
-                line_number_map["pc_raw_ast_map"][pc] = ast_node
-
-        if item.error_msg is not None:
-            line_number_map["error_map"][pc] = item.error_msg
-
-    note_breakpoint(line_number_map, pc, item)
-
-
-# NOTE: this is dead code, we don't emit DEBUG anymore.
-def note_breakpoint(line_number_map, pc, item):
-    # Record line number attached to pc
-    if item == "DEBUG":
-        # Is PC debugger, create PC breakpoint.
-        if item.pc_debugger:
-            line_number_map["pc_breakpoints"].add(pc)
-        # Create line number breakpoint.
-        else:
-            line_number_map["breakpoints"].add(item.lineno + 1)
-
-
-##############################
-# assembly optimizer
-##############################
-
-_TERMINAL_OPS = ("JUMP", "RETURN", "REVERT", "STOP", "INVALID")
-
-
-def _prune_unreachable_code(assembly):
-    # delete code between terminal ops and JUMPDESTS as those are
-    # unreachable
-    changed = False
-    i = 0
-    while i < len(assembly) - 1:
-        if assembly[i] in _TERMINAL_OPS:
-            # find the next jumpdest or data section
-            for j in range(i + 1, len(assembly)):
-                next_is_reachable = isinstance(assembly[j], (Label, DataHeader))
-                if next_is_reachable:
-                    break
-            else:
-                # fixup an off-by-one if we made it to the end of the assembly
-                # without finding an jumpdest or sublist
-                j = len(assembly)
-            changed |= j > i + 1
-            del assembly[i + 1 : j]
-
-        i += 1
-
-    return changed
-
-
-def _prune_inefficient_jumps(assembly):
-    # prune sequences `PUSHLABEL x JUMP LABEL x` to `LABEL x`
-    changed = False
-    i = 0
-    while i < len(assembly) - 2:
-        if (
-            isinstance(assembly[i], PUSHLABEL)
-            and assembly[i + 1] == "JUMP"
-            and is_label(assembly[i + 2])
-            and assembly[i + 2] == assembly[i].label
-        ):
-            # delete PUSHLABEL x JUMP
-            changed = True
-            del assembly[i : i + 2]
-        else:
-            i += 1
-
-    return changed
-
-
-def _optimize_inefficient_jumps(assembly):
-    # optimize sequences
-    # `PUSHLABEL common JUMPI PUSHLABEL x JUMP LABEL common`
-    # to `ISZERO PUSHLABEL x JUMPI LABEL common`
-    changed = False
-    i = 0
-    while i < len(assembly) - 4:
-        if (
-            isinstance(assembly[i], PUSHLABEL)
-            and assembly[i + 1] == "JUMPI"
-            and isinstance(assembly[i + 2], PUSHLABEL)
-            and assembly[i + 3] == "JUMP"
-            and isinstance(assembly[i + 4], Label)
-            and assembly[i].label == assembly[i + 4]
-        ):
-            changed = True
-            assembly[i] = "ISZERO"
-            assembly[i + 1] = assembly[i + 2]
-            assembly[i + 2] = "JUMPI"
-            del assembly[i + 3 : i + 4]
-        else:
-            i += 1
-
-    return changed
-
-
-def _merge_jumpdests(assembly):
-    # When we have multiple JUMPDESTs in a row, or when a JUMPDEST
-    # is immediately followed by another JUMP, we can skip the
-    # intermediate jumps.
-    # (Usually a chain of JUMPs is created by a nested block,
-    # or some nested if statements.)
-    changed = False
-    i = 0
-    while i < len(assembly) - 2:
-        if is_label(assembly[i]):
-            current_symbol = assembly[i]
-            if is_label(assembly[i + 1]):
-                # LABEL x LABEL y
-                # replace all instances of PUSHLABEL x with PUSHLABEL y
-                # (could also remove PUSH_OFST and DATA_ITEM, but doesn't
-                #  affect correctness)
-                new_symbol = assembly[i + 1]
-                if new_symbol != current_symbol:
-                    for j in range(len(assembly)):
-                        if (
-                            isinstance(assembly[j], PUSHLABEL)
-                            and assembly[j].label == current_symbol
-                        ):
-                            assembly[j].label = new_symbol
-                            changed = True
-            elif isinstance(assembly[i + 1], PUSHLABEL) and assembly[i + 2] == "JUMP":
-                # LABEL x PUSHLABEL y JUMP
-                # replace all instances of PUSHLABEL x with PUSHLABEL y
-                # (could also remove PUSH_OFST and DATA_ITEM, but doesn't
-                #  affect correctness)
-                new_symbol = assembly[i + 1].label
-                for j in range(len(assembly)):
-                    if isinstance(assembly[j], PUSHLABEL) and assembly[j].label == current_symbol:
-                        assembly[j].label = new_symbol
-                        changed = True
-
-        i += 1
-
-    return changed
-
-
-_RETURNS_ZERO_OR_ONE = {
-    "LT",
-    "GT",
-    "SLT",
-    "SGT",
-    "EQ",
-    "ISZERO",
-    "CALL",
-    "STATICCALL",
-    "CALLCODE",
-    "DELEGATECALL",
-}
-
-
-def _merge_iszero(assembly):
-    changed = False
-
-    i = 0
-    # list of opcodes that return 0 or 1
-    while i < len(assembly) - 2:
-        if (
-            isinstance(assembly[i], str)
-            and assembly[i] in _RETURNS_ZERO_OR_ONE
-            and assembly[i + 1 : i + 3] == ["ISZERO", "ISZERO"]
-        ):
-            changed = True
-            # drop the extra iszeros
-            del assembly[i + 1 : i + 3]
-        else:
-            i += 1
-    i = 0
-    while i < len(assembly) - 3:
-        # ISZERO ISZERO could map truthy to 1,
-        # but it could also just be a no-op before JUMPI.
-        if (
-            assembly[i : i + 2] == ["ISZERO", "ISZERO"]
-            and isinstance(assembly[i + 2], PUSHLABEL)
-            and assembly[i + 3] == "JUMPI"
-        ):
-            changed = True
-            del assembly[i : i + 2]
-        else:
-            i += 1
-
-    return changed
-
-
-def _prune_unused_jumpdests(assembly):
-    changed = False
-
-    used_jumpdests: set[Label] = set()
-
-    # find all used jumpdests
-    for item in assembly:
-        if isinstance(item, PUSHLABEL):
-            used_jumpdests.add(item.label)
-
-        if isinstance(item, PUSH_OFST) and isinstance(item.label, Label):
-            used_jumpdests.add(item.label)
-
-        if isinstance(item, DATA_ITEM) and isinstance(item.data, Label):
-            # add symbols used in data sections as they are likely
-            # used for a jumptable.
-            used_jumpdests.add(item.data)
-
-    # delete jumpdests that aren't used
-    i = 0
-    while i < len(assembly):
-        if is_label(assembly[i]) and assembly[i] not in used_jumpdests:
-            changed = True
-            del assembly[i]
-        else:
-            i += 1
-
-    return changed
-
-
-def _stack_peephole_opts(assembly):
-    changed = False
-    i = 0
-    while i < len(assembly) - 2:
-        if assembly[i : i + 3] == ["DUP1", "SWAP2", "SWAP1"]:
-            changed = True
-            del assembly[i + 2]
-            assembly[i] = "SWAP1"
-            assembly[i + 1] = "DUP2"
-            continue
-        # usually generated by with statements that return their input like
-        # (with x (...x))
-        if assembly[i : i + 3] == ["DUP1", "SWAP1", "POP"]:
-            # DUP1 SWAP1 POP == no-op
-            changed = True
-            del assembly[i : i + 3]
-            continue
-        # usually generated by nested with statements that don't return like
-        # (with x (with y ...))
-        if assembly[i : i + 3] == ["SWAP1", "POP", "POP"]:
-            # SWAP1 POP POP == POP POP
-            changed = True
-            del assembly[i]
-            continue
-        if (
-            isinstance(assembly[i], str)
-            and assembly[i].startswith("SWAP")
-            and assembly[i] == assembly[i + 1]
-        ):
-            changed = True
-            del assembly[i : i + 2]
-        if assembly[i] == "SWAP1" and str(assembly[i + 1]).lower() in COMMUTATIVE_OPS:
-            changed = True
-            del assembly[i]
-        if assembly[i] == "DUP1" and assembly[i + 1] == "SWAP1":
-            changed = True
-            del assembly[i + 1]
-        i += 1
-
-    return changed
-
-
-# optimize assembly, in place
-def optimize_assembly(assembly):
-    for _ in range(1024):
-        changed = False
-
-        changed |= _prune_unreachable_code(assembly)
-        changed |= _merge_iszero(assembly)
-        changed |= _merge_jumpdests(assembly)
-        changed |= _prune_inefficient_jumps(assembly)
-        changed |= _optimize_inefficient_jumps(assembly)
-        changed |= _prune_unused_jumpdests(assembly)
-        changed |= _stack_peephole_opts(assembly)
-
-        if not changed:
-            return
-
-    raise CompilerPanic("infinite loop detected during assembly reduction")  # pragma: nocover
-
-
-SYMBOL_SIZE = 2  # size of a PUSH instruction for a code symbol
-
-
-# predict what length of an assembly [data] node will be in bytecode
-def get_data_segment_lengths(assembly: list[AssemblyInstruction]) -> list[int]:
-    ret = []
-    for item in assembly:
-        if isinstance(item, DataHeader):
-            ret.append(0)
-            continue
-        if len(ret) == 0:
-            # haven't yet seen a data header
-            continue
-        assert isinstance(item, DATA_ITEM)
-        if is_label(item.data):
-            ret[-1] += SYMBOL_SIZE
-        elif isinstance(item.data, bytes):
-            ret[-1] += len(item.data)
-        else:  # pragma: nocover
-            raise ValueError(f"invalid data {type(item)} {item}")
-
-    return ret
-
-
-##############################
-# assembly to evm bytecode
-##############################
-
-
-def _compile_data_item(item: DATA_ITEM, symbol_map: dict[Label, int]) -> bytes:
-    if isinstance(item.data, bytes):
-        return item.data
-    if isinstance(item.data, Label):
-        symbolbytes = symbol_map[item.data].to_bytes(SYMBOL_SIZE, "big")
-        return symbolbytes
-
-    raise CompilerPanic(f"Invalid data {type(item.data)}, {item.data}")  # pragma: nocover
-
-
-T = TypeVar("T")
-
-
-def _add_to_symbol_map(symbol_map: dict[T, int], item: T, value: int):
-    if item in symbol_map:  # pragma: nocover
-        raise CompilerPanic(f"duplicate label: {item}")
-    symbol_map[item] = value
-
-
-def assembly_to_evm(assembly: list[AssemblyInstruction]) -> tuple[bytes, dict[str, Any]]:
-    """
-    Generate bytecode and source map from assembly
-
-    Returns:
-        bytecode: bytestring of the EVM bytecode
-        source_map: source map dict that gets output for the user
-    """
-    # This API might seem a bit strange, but it's backwards compatible
-    symbol_map, const_map, source_map = resolve_symbols(assembly)
-    bytecode = _assembly_to_evm(assembly, symbol_map, const_map)
-    return bytecode, source_map
-
-
-# resolve symbols in assembly
-def resolve_symbols(
-    assembly: list[AssemblyInstruction],
-) -> tuple[dict[Label, int], dict[CONSTREF, int], dict[str, Any]]:
-    """
-    Construct symbol map from assembly list
-
-    Returns:
-        symbol_map: dict from labels to values
-        const_map: dict from CONSTREFs to values
-        source_map: source map dict that gets output for the user
-    """
-    source_map: dict[str, Any] = {
-        "breakpoints": OrderedSet(),
-        "pc_breakpoints": OrderedSet(),
-        "pc_jump_map": {0: "-"},
-        "pc_raw_ast_map": {},
-        "error_map": {},
-    }
-
-    symbol_map: dict[Label, int] = {}
-    const_map: dict[CONSTREF, int] = {}
-
-    pc: int = 0
-
-    # resolve constants
-    for item in assembly:
-        if isinstance(item, CONST):
-            # should this be merged into the symbol map?
-            _add_to_symbol_map(const_map, CONSTREF(item.name), item.value)
-
-    # resolve labels (i.e. JUMPDEST locations) to actual code locations,
-    # and simultaneously build the source map.
-    for i, item in enumerate(assembly):
-        # add it to the source map
-        note_line_num(source_map, pc, item)
-
-        # update pc_jump_map
-        if item == "JUMP":
-            assert i != 0  # otherwise we can get assembly[-1]
-            last = assembly[i - 1]
-            if isinstance(last, PUSHLABEL) and last.label.label.startswith("internal"):
-                if last.label.label.endswith("cleanup"):
-                    # exit an internal function
-                    source_map["pc_jump_map"][pc] = "o"
-                else:
-                    # enter an internal function
-                    source_map["pc_jump_map"][pc] = "i"
-            else:
-                # everything else
-                source_map["pc_jump_map"][pc] = "-"
-        elif item in ("JUMPI", "JUMPDEST"):
-            source_map["pc_jump_map"][pc] = "-"
-
-        if item == "DEBUG":
-            continue  # "debug" opcode does not go into bytecode
-
-        if isinstance(item, CONST):
-            continue  # CONST declarations do not go into bytecode
-
-        # update pc
-        if isinstance(item, Label):
-            _add_to_symbol_map(symbol_map, item, pc)
-            pc += 1  # jumpdest
-
-        elif isinstance(item, DataHeader):
-            # Don't increment pc as the symbol itself doesn't go into code
-            _add_to_symbol_map(symbol_map, item.label, pc)
-
-        elif isinstance(item, PUSHLABEL):
-            pc += SYMBOL_SIZE + 1  # PUSH2 highbits lowbits
-
-        elif isinstance(item, PUSH_OFST):
-            assert isinstance(item.ofst, int), item
-            # [PUSH_OFST, (Label foo), bar] -> PUSH2 (foo+bar)
-            # [PUSH_OFST, _mem_foo, bar] -> PUSHN (foo+bar)
-            if isinstance(item.label, Label):
-                pc += SYMBOL_SIZE + 1  # PUSH2 highbits lowbits
-            elif isinstance(item.label, CONSTREF):
-                const = const_map[item.label]
-                val = const + item.ofst
-                pc += calc_push_size(val)
-            else:  # pragma: nocover
-                raise CompilerPanic(f"invalid ofst {item.label}")
-
-        elif isinstance(item, DATA_ITEM):
-            if isinstance(item.data, Label):
-                pc += SYMBOL_SIZE
-            else:
-                assert isinstance(item.data, bytes)
-                pc += len(item.data)
-        elif isinstance(item, int):
-            assert 0 <= item < 256
-            pc += 1
-        else:
-            assert isinstance(item, str) and item in get_opcodes(), item
-            pc += 1
-
-    source_map["breakpoints"] = list(source_map["breakpoints"])
-    source_map["pc_breakpoints"] = list(source_map["pc_breakpoints"])
-
-    # magic -- probably the assembler should actually add this label
-    _add_to_symbol_map(symbol_map, Label("code_end"), pc)
-
-    return symbol_map, const_map, source_map
-
-
-# helper function
-def _compile_push_instruction(assembly: list[AssemblyInstruction]) -> bytes:
-    push_mnemonic = assembly[0]
-    assert isinstance(push_mnemonic, str) and push_mnemonic.startswith("PUSH")
-    push_instr = PUSH_OFFSET + int(push_mnemonic[4:])
-    ret = [push_instr]
-
-    for item in assembly[1:]:
-        assert isinstance(item, int)
-        ret.append(item)
-    return bytes(ret)
-
-
-def _assembly_to_evm(
-    assembly: list[AssemblyInstruction],
-    symbol_map: dict[Label, int],
-    const_map: dict[CONSTREF, int],
-) -> bytes:
-    """
-    Assembles assembly into EVM bytecode
-
-    Parameters:
-        assembly: list of asm instructions
-        symbol_map: dict from labels to resolved locations in the code
-        const_map: dict from constrefs to their values
-
-    Returns: bytes representing the bytecode
-    """
-    ret = bytearray()
-
-    # now that all symbols have been resolved, generate bytecode
-    # using the symbol map
-    for item in assembly:
-        if item in ("DEBUG",):
-            continue  # skippable opcodes
-        elif isinstance(item, CONST):
-            continue  # CONST things do not show up in bytecode
-        elif isinstance(item, DataHeader):
-            continue  # DataHeader does not show up in bytecode
-
-        elif isinstance(item, PUSHLABEL):
-            # push a symbol to stack
-            label = item.label
-            bytecode = _compile_push_instruction(PUSH_N(symbol_map[label], n=SYMBOL_SIZE))
-            ret.extend(bytecode)
-
-        elif isinstance(item, Label):
-            jumpdest_opcode = get_opcodes()["JUMPDEST"][0]
-            assert jumpdest_opcode is not None  # help mypy
-            ret.append(jumpdest_opcode)
-
-        elif isinstance(item, PUSH_OFST):
-            # PUSH_OFST (LABEL foo) 32
-            # PUSH_OFST (const foo) 32
-            if isinstance(item.label, Label):
-                ofst = symbol_map[item.label] + item.ofst
-                bytecode = _compile_push_instruction(PUSH_N(ofst, SYMBOL_SIZE))
-            else:
-                assert isinstance(item.label, CONSTREF)
-                ofst = const_map[item.label] + item.ofst
-                bytecode = _compile_push_instruction(PUSH(ofst))
-
-            ret.extend(bytecode)
-
-        elif isinstance(item, int):
-            ret.append(item)
-        elif isinstance(item, str) and item.upper() in get_opcodes():
-            opcode = get_opcodes()[item.upper()][0]
-            # TODO: fix signature of get_opcodes()
-            assert opcode is not None  # help mypy
-            ret.append(opcode)
-        elif isinstance(item, DATA_ITEM):
-            ret.extend(_compile_data_item(item, symbol_map))
-        elif item[:4] == "PUSH":
-            ret.append(PUSH_OFFSET + int(item[4:]))
-        elif item[:3] == "DUP":
-            ret.append(DUP_OFFSET + int(item[3:]))
-        elif item[:4] == "SWAP":
-            ret.append(SWAP_OFFSET + int(item[4:]))
-        else:  # pragma: no cover
-            # unreachable
-            raise ValueError(f"Weird symbol in assembly: {type(item)} {item}")
-
-    return bytes(ret)
