+++ vyper/venom/venom_to_assembly.py
@@ -2,14 +2,13 @@ from __future__ import annotations
 
 from typing import Any, Iterable
 
+from build.lib.vyper.ir.compile_ir import DataHeader
+from vyper.evm.assembler.instructions import DATA_ITEM, PUSH
 from vyper.exceptions import CompilerPanic, StackTooDeep
 from vyper.ir.compile_ir import (
-    DATA_ITEM,
-    PUSH,
     PUSH_OFST,
     PUSHLABEL,
     AssemblyInstruction,
-    DataHeader,
     Label,
     TaggedInstruction,
     optimize_assembly,
